LayoutTests/fast/page-color-sampling/blend-modal-background-color-expected.txt

PASS colors.top is "rgb(51, 51, 51)"
PASS successfullyParsed is true

TEST COMPLETE
